fix tests and simplify eventsources

+++ Stack/Opc.Ua.Core/Stack/Tcp/TcpListenerChannel.cs
@@ -143,7 +143,7 @@ namespace Opc.Ua.Bindings
                     return;
                 }
 
-                Utils.EventLog.SendResponse(ChannelId, requestId);
+                Utils.EventLog.SendResponse((int)ChannelId, (int)requestId);
 
                 BufferCollection buffers = null;
